+++ packages/ui/src/components/SupportScreen/SupportScreenStyles.ts
@@ -1,0 +1,11 @@
+import {createStyles, makeStyles} from "@material-ui/core/styles";
+
+const SupportScreenStyles = makeStyles(() =>
+  createStyles({
+    window: {
+      height: "70vh",
+    },
+  }),
+);
+
+export default SupportScreenStyles;