+++ packages/web/pages_/support.tsx
@@ -4,7 +4,6 @@ import * as React from "react";
 
 import AuthContext from "@sentrei/common/context/AuthContext";
 import {analytics} from "@sentrei/common/utils/firebase";
-import Footer from "@sentrei/ui/components/Footer";
 import Loader from "@sentrei/ui/components/Loader";
 import SupportScreen from "@sentrei/ui/components/SupportScreen";
 import SentreiHeader from "@sentrei/web/components/SentreiHeader";
@@ -26,9 +25,8 @@ const Support: NextPage = () => {
 
   return (
     <>
-      <SentreiHeader />
+      <SentreiHeader papercups={false} />
       <SupportScreen />
-      <Footer />
     </>
   );
 };